+++ src/audio/dcblock/CMakeLists.txt
@@ -2,3 +2,9 @@ add_local_sources(sof dcblock.c)
 add_local_sources(sof dcblock_generic.c)
 add_local_sources(sof dcblock_hifi3.c)
 add_local_sources(sof dcblock_hifi4.c)
+
+if(CONFIG_IPC_MAJOR_3)
+	add_local_sources(sof dcblock_ipc3.c)
+elseif(CONFIG_IPC_MAJOR_4)
+	add_local_sources(sof dcblock_ipc4.c)
+endif()
